create samples and migrate_templates under templates/ directory to separate logic
