+++ lib/dradis/plugins.rb
@@ -96,6 +96,7 @@ require 'dradis/plugins/configurable'
 require 'dradis/plugins/settings'
 require 'dradis/plugins/settings/adapters/db'
 require 'dradis/plugins/settings/adapters/encrypted_configuration'
-require 'dradis/plugins/templates'
+require 'dradis/plugins/templates/samples'
+require 'dradis/plugins/templates/migrate_templates'
 require 'dradis/plugins/thor'
 require 'dradis/plugins/thor_helper'
